+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ in development
   in default etc. (improvement)
 * Fix action chain so it doesn't end up in an infinite loop if an action which is part of the chain
   is canceled. (bug fix)
+* Allow jinja templating to be used in ``message`` and ``data`` field for notifications.(new feature)
 * Add tools for purging executions (also, liveactions with it) and trigger instances older than
   certain UTC timestamp from the db in bulk.